+++ src/main/java/site/siredvin/progressiveperipherals/common/configuration/ProgressivePeripheralsConfig.java
@@ -22,6 +22,7 @@ public class ProgressivePeripheralsConfig {
     public static boolean enableStatueWorkbench;
     public static boolean enableAbstractiumPedestal;
     public static boolean enableRecipeRegistry;
+    public static boolean enableEventDistributor;
     public static boolean enableScientificAutomataCore;
     public static boolean enableEnchantingAutomataCore;
     public static boolean enableSmithingAutomataCore;
@@ -57,6 +58,7 @@ public class ProgressivePeripheralsConfig {
         final ForgeConfigSpec.BooleanValue ENABLE_STATUE_WORKBENCH;
         final ForgeConfigSpec.BooleanValue ENABLE_ABSTRACTIUM_PEDESTAL;
         final ForgeConfigSpec.BooleanValue ENABLE_RECIPE_REGISTRY;
+        final ForgeConfigSpec.BooleanValue ENABLE_EVENT_DISTRIBUTOR;
         final ForgeConfigSpec.BooleanValue ENABLE_ENCHATING_AUTOMATA_CORE;
         final ForgeConfigSpec.BooleanValue ENABLE_SMITHING_AUTOMATA_CORE;
         final ForgeConfigSpec.BooleanValue ENABLE_BREWING_AUTOMATA_CORE;
@@ -100,6 +102,7 @@ public class ProgressivePeripheralsConfig {
             ENABLE_STATUE_WORKBENCH = builder.define("enableStatueWorkbench", true);
             ENABLE_ABSTRACTIUM_PEDESTAL = builder.define("enableAbstractiumPedestal", true);
             ENABLE_RECIPE_REGISTRY = builder.define("enableRecipeRegistry", true);
+            ENABLE_EVENT_DISTRIBUTOR = builder.define("enableEventDistributor", true);
             ENABLE_ENCHATING_AUTOMATA_CORE = builder.define("enableEnchantingAutomataCore", true);
             ENABLE_SMITHING_AUTOMATA_CORE = builder.define("enableSmithingAutomataCore", true);
             ENABLE_BREWING_AUTOMATA_CORE = builder.define("enableBrewingAutomataCore", true);