+++ src/main/java/site/siredvin/progressiveperipherals/integrations/computercraft/peripherals/FreeOperation.java
@@ -9,7 +9,8 @@ import java.util.Map;
 public enum FreeOperation implements IPeripheralOperation<Object> {
     EXTRACT(5_000),
     QUERY_REGISTRY(5_000),
-    DUPLICATE_ITEM(1_000);
+    DUPLICATE_ITEM(1_000),
+    QUEUE_EVENT(100);
 
     private ForgeConfigSpec.IntValue cooldown;
     private final int defaultCooldown;